pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "kedro-pypi-monitor"
description = "Monitor for PyPI downloads of Kedro"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "google-cloud-bigquery",
  "google-cloud-bigquery-storage",
  "kedro>=0.19.6",
  "kedro-datasets[polars]",
  "polars>=1.0",
  "pyarrow",
]
authors = [
    {name = "Juan Luis Cano Rodríguez", email = "juan_luis_cano@mckinsey.com"},
]
dynamic = ["version"]

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"

[tool.kedro]
project_name = "kedro-pypi-monitor"
package_name = "kedro_pypi_monitor"
kedro_init_version = "0.19.6"

[tool.ruff]
show-fixes = true

[tool.ruff.lint]
select = [
    "F",  # Pyflakes
    "E",  # Pycodestyle
    "W",  # Pycodestyle
    "UP",  # pyupgrade
    "I",  # isort
    "PL",  # Pylint
]
